requirements.txt
PyQt5
langchain
openai
PyPDF2
tiktoken
faiss-cpu
python-docx
